:lipstick: Style refactor
