--- src/containers/Counter/Base.js
@@ -1,10 +1,0 @@
-import styled from 'styled-components';
-
-const Base = styled.div`
-  margin: 0 auto;
-  width: 100%;
-  max-width: 768px;
-  text-align: center;
-`;
-
-export default Base;
