+++ src/containers/NotFound/index.js
@@ -1,11 +1,11 @@
 import React from 'react';
-import Wrapper from './Wrapper';
+import Container from 'components/Container';
 import H2 from './H2';
 
 const NotFound = () => (
-  <Wrapper>
+  <Container center>
     <H2>Page Not Found!</H2>
-  </Wrapper>
+  </Container>
 );
 
 export default NotFound;
