--- src/containers/Todo/Wrapper.js
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-
-const Wrapper = styled.div`
-  margin: 0 auto;
-  width: 100%;
-  max-width: 768px;
-`;
-
-export default Wrapper;
